Backend/requirements.txt
fastapi>=0.110
uvicorn[standard]
sqlalchemy>=2.0
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt<5
pandas
openpyxl
python-multipart
email-validator
